Final/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.3.0
filetype==1.2.0
flatbuffers==25.2.10
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jh2==5.0.9
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-community==0.3.26
langchain-core==0.3.66
langchain-nvidia-ai-endpoints==0.3.10
langchain-text-splitters==0.3.8
langsmith==0.4.1
marshmallow==3.26.1
multidict==6.5.0
mypy_extensions==1.1.0
niquests==3.14.1
numpy==2.2.6
openai==1.91.0
openmeteo_requests==1.5.0
openmeteo_sdk==1.20.1
orjson==3.10.18
packaging==24.2
pandas==2.3.0
propcache==0.3.2
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
qh3==1.5.3
requests==2.32.4
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
tenacity==9.1.2
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
urllib3-future==2.13.900
wassima==1.2.2
yarl==1.20.1
zstandard==0.23.0
chromadb